cookiejar: Prevent updates to cookiejar when another operation is ongoing
When cookies are added/removed some listeners might change their own
cookiejars and then try to add the same cookie again to the cookiejar
again and we would have an infinite recursion.

This would be prevented by checking that the cookies are the same
and gstcookiejar would ignore this add/removal. It could happen,
however, that an implementation of a cookie has a missing field or
uses a less precise 'expires' attribute that would lead to the
cookie comparison thinking the cookie is different.